Cleanup crap

+++ src/coreclr/jit/hwintrinsiccodegenxarch.cpp
@@ -827,7 +827,7 @@ void CodeGen::GenHWIntrinsicJumpTableFallback(NamedIntrinsic            intrinsi
     assert(!HWIntrinsicInfo::IsAvx2GatherIntrinsic(intrinsic));
     Emitter& emit = *GetEmitter();
 
-    const unsigned maxByte = static_cast<unsigned>(HWIntrinsicInfo::GetImmOpUpperBound(intrinsic) + 1);
+    const unsigned maxByte = HWIntrinsicInfo::GetImmOpUpperBound(intrinsic) + 1;
     assert(maxByte <= 256);
 
     insGroup** labels;
